info: fix listing door interfaces for multihomed nodes
As accurately described in ticket RT#6198, there is a problem with 
the info-service if a door has multiple interfaces that resolve to 
the same name then a single interface is published rather than one 
for each interface.  If the published interface happens not to be the 
first in the list then the info-provider will fail to publish a valid 
URL for this door.

This patch addresses this by assigning each interface a unique ID.  
Currently, the ID is built by concatenating the interface name (which 
is possiblely non-unique) with the index of the interface.  The 
combination is unique.

Ticket: http://rt.dcache.org/Ticket/Display.html?id=6198
Patch: http://rb.dcache.org/r/3266/
Acked-by: Gerd Behrmann
Target: trunk
Request: 1.9.12
Request: 1.9.11
Request: 1.9.10