PinManager: Perform pool selection in PinManager
When pinning a file, PinManager asks PoolManager to select a read pool.
This is bad for several reasons:

- PoolManager updates its cost matrix according to the selection, even
  though pin manager is not actually going to read the file. When the
  file is finally read the cost matrix is updated again.

- PoolManager perform cost triggered pool to pool selections, even though
  the cost may have changed dramatically by the time the file is
  actually accessed.

- PoolManager may choose to return different read pools for each
  request, resulting in the file being pinned on several pools.

- The request introduces latency, increasing the time it takes to
  prepeare a get TURL.

- The PinManager puts load on PoolManager, a component that may eventually
  become a bottleneck in dCache.


This patch changes the relationship between PinManager and PoolManager.
The patch attempts to avoid the roundtrip to the PoolManager for the
most common pinning operation: Pinning a file already present on a read
pool. The patch achieves this by periodically asking PoolManager to
serialize its PoolMonitorV5 instance and send it to PinManager.
PoolMonitorV5 embeds information about the pool selection unit, the
current pool status and cost (although cost is not needed by PinManager
right now). PinManager then attempts to select a pool on which to pin
the file based on that information. Should this fail, maybe because the
file needs to be staged or copied to a read pool, PinManager falls back
to asking PoolManager as it did before.

The logic in PinRequestProcessor.selectReadPool is a modified copy of
the logic present in RequestContainerV5.askIfAvailable. We copy the
logic of minCostCut to select a pool deterministically.

The current implementation of how PoolMonitorV5 is exported is very much
PinManager specific. We may eventually choose to generalize this so that
doors may use the same approach. I envision that we introduce an
interface for selecting a pool for various purposes. PoolManager would
publish an implementation of that interface to the doors. PoolManager
could then choose among several possible implementations of that
interface: One could always ask PoolManager, one could perform common
operations in the door and fallback to PoolManager. We could even let
SpaceManager wrap this object and inject its own lookup code into what
is passed on to the door.

PinManager currently polls PoolManager periodically for a copy of
PoolMonitorV5. When we generalize this infrastructure as outlined above
then we should switch from polling to a publish subscribe model: Doors
should subscribe to updates from PoolManager. This is in particular
relevant for doors, as doors will need the current cost. A publish
subscribe model would allow PoolManager to adjust the publishing rate
when the cost changes rapidly.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3195/