pnfsmanger: Move permission check into chimera name space provider
Moves the permission check previously performed in
PermissionHandlerNameSpaceProvider into ChimeraNameSpaceProvider. This
allows us to reduce the overhead of those checks by avoiding repeated
reads of the same attributes.

PermissionHandlerNameSpaceProvider is still used for PNFS:
PnfsNameSpaceProvider is introduced as a helper class to allow the two
providers to have the same interface.

There is room for further improvements in some of the checks.

This patch provides the foundation for another patch that adds an
optional lookup permission check for the complete path and in the future
for dynamic ACL inheritance.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3214/