gplazma: Let gridmap plugin map to UserNamePrincipal
The gridmap plugin currently maps to GroupNamePrincipal. This was done
as a quick hack to make it chainable with the vorolemap and authzdb
plugins.

A better solution is to let gridmap map to UserNamePrincipal and let the
authzdb plugin process both user names and group names. Since the
storage-authzdb file format doesn't distinguish between users and groups
it needs to prioritize the possible candidates for UID and primary GID.

Rather than hardcoding an order the module introduces two new
configuration parameters: uid and gid. These determine the order in
which candidate mappings are processed for the UID and the primary GID.
Values are a comma separated list of the key words uid,gid,login,user,
and group. uid is the login uid (as used with DCAP) and gid is the login
gid (as used with DCAP). uid only has an effect for uid and gid for gid.
login refers to the login name (as used with DCAP and FTP); user refers
to the user name (as generated by gridmap or through password based
authentication) and group refers to the primary group name (as generated
by vorolemap).

The default for uid is uid,login,user,group and the default for gid is
gid,login,group,user. Ie for UID if the DCAP client specified a uid then
it is used. Otherwise if a login name was specified then it is used.
Otherwise if a user name is available (typically from gridmap) then it
is used. Otherwise if a primary group name is available (typically from
vorolemap) then is is used. The process for GID is similar.

With the default values chaining vorolemap, gridmap and authzdb will
cause the UID to be determined by the gridmap mapping and the primary
GID by the vorolemap mapping (if available).

Secondary GIDs are always included from all available mappings. This is
different from what would have happened before or what happened in
gPlazma 1. Including additional secondary GIDs that the user is in fact
a member of doesn't have any negative side effects as far as I can see.

The vorolemap and gridmap plugins no longer add
GroupNamePrincipal/UserNamePrincipal to the authorized principals sets.
Instead the authzdb plugin adds the user name that was used for the
mapping from which the UID was taken.

LoginGidPrincipal was changed to not implement GroupPrincipal: The
primary group attribute doesn't make sense for this principal.

Target: trunk
Request: 1.9.12
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3269/