chimera: Add support for resolving a path to all inodes of the path
Chimera has a method, path2inode, to resolve a path into a single inode.
The inode is the inode of the file or directory the path identifies.

This patch adds the path2inodes method. It too resolves the path, but it
returns the inodes of all path elements, not just the final inode. This
is quite useful when checking lookup permissions, as we need the inodes
of the path elements too.

The stat cache of the inode is prepulated. Since we have to access the
record in t_inodes anyway, returning the data to the client adds minimal
overhead and avoids that we have to fetch the stat data for the inodes
afterwards.

For postgresql a stored procedure resolves the complete path and returns
a set of inodes in a single roundtrib to the database.

The pathToInodes method in the chimera name space provider is updated to
utilize the new functionality. In the current revision we still only
evaluate the lookup permissions for the parent of the target object.
Even with this minimal change we save several round trips to the
database. In a future update we can extend the permission check to
enforce lookup permission checks for the complete path.

RELEASE NOTES: Chimera users must recreate the stored procedures by
applying pgsql-procedures.sql.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3216/