httpd: Fix redundant null check
Findbugs warned about a guaranteed NPE for a branch in the pool manager
view.

The warning is bogus since the two arrays are newer null in the cases in
which the comparator is used. The code is however bogus: If the
arguments where null then the code would surely fail with NPE and hence
there is no reason to check for a null value.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3288/