admin: Avoid NPE warning
The ssh engine instantiates the Md5 class protected by a try-catch; the
catch catches Exception and 'recovers' from the error by setting the
reference that should have been initialized to null. Right after the
try-catch the reference is dereferences. If an exception would be thrown
during construction of Md5 an NPE would follow.

Analysing the Md5 constructor shows that no exception can be thrown and
thus the try-catch is redundant. The patch removes it.

Thanks to findbugs.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3289/