cells: Make shutdown sequence more robust
Upon OOM failures dCache automatically restarts. We have observed that
the restart sometimes hangs. One theory about why this happens is that
the shutdown itself fails with an OOM error.

This patch tries to address that problem: It removes a number temporary
objects and preallocates others. The shutdown sequence is however so
long and complex that we cannot get rid of all allocations. The patch
uses the common technique of preallocating a large (2MB) block of memory
which is released upon OOM. This increases the chance that the shutdown
succeeds.

If this isn't enough to resolve the problem we can add additional checks
and call Runtime.halt() when we detect repeated OOM failures.

Target: trunk
Request: 1.9.12
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3291/