utils: add NetworkInterfaceView
Add a serializable snapshot of a NetworkInterface object.  The 
NetworkInterfaceView provides a serializable object allowing remote 
components to know about the interfaces of a door.

Unfortunately, InterfaceAddress is also not serialisable and has no 
accessible constructor, so the patch also provides a 
InterfaceAddressView class that is a serialisable version of 
InterfaceAddress.

The NetworkInterfaceView object will walk down the parent-child tree 
to record information about all sub-interfaces but does not follow the 
parent link.  This places a requirement that the source 
NetworkInterface object is the root of the parent-child tree.

The patch is intended to provide a mechanism to allow doors to relay 
more detailed information about their interface to the LoginBroker, 
which can then provide more detailed information about the doors 
available in dCache.

Target: trunk
Patch: http://rb.dcache.org/r/2943/
Acked-by: Gerd Behrmann