unit-tests: fix recently introduced tests
The recently committed patch introduced extra tests that, when building 
the fixture, rely on JUnit's assume* static methods.

Unfortunately, these methods are first available in a newer version of 
JUnit that we are currently using; so committing this patch broke our 
unit testing of trunk.

It turns out that upgrading the version of JUnit trigger problems with 
our existing tests.  These problems must be addressed, but that will be 
the subject of a future patch; however, this patch aims to allow us to 
build and unit-test trunk cleanly.

Target: trunk
Patch: http://rb.dcache.org/r/3296
Acked-by: Tigran Mkrtchyan