dcap: Use ConcurrentHashMap for session map
The session map is accessed from both the message thread and from the
DCAP communication thread. In most cases access is synchronized on the
interpreter instance, however in some cases (eg the removeUs() method)
the access is missing.

This patch replaces the HashMap with a ConcurrentHashMap.

Signed-off-by: Gerd Behrmann
Patch: http://rb.dcache.org/r/3300/