packaging: perserve billing directory ownership across RPM upgrade
We describe how it is possible to run dCache as non-root user.  There 
are certain directories that must be writeable for whichever user 
dCache runs as.  One of these directories is the directory under which 
billing records are stored, assuming the node is running the billing 
cell.

Unfortunately, installing or upgrading the dCache RPM will result in 
the billing directory reverting ownership to root.  This will cause 
the next attempt to start dCache to fail, forcing the sysadmin to 
manually adjust the permissions of the billing directory after each 
upgrade.

This patch fixes the problem by introducing the idea of preserving 
some state across an RPM upgrade.  This is achieved by creating a 
script, prior to the upgrade, that will reestablish the desired state 
after the upgrade.  The script is then envoked after the new version's 
files have been installed.  Since there is no shared state between the 
%pre and %post scriptlet invocations, we are forced to place the 
state-reestablishing script in a standard location.

Currently, only the ownership of the billing directory is preserved.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3338/
Acked-by: Gerd Behrmann
Acked-by: Tigran Mkrtchyan