pnfs-manager: fix getFileAttributes for root directory
In trunk, the getFileAttributes method from Chimera namespace provider 
now checks that the user is allowed to read the requested information 
by checking the permissions for the entire path.  Checking permissions 
for the root directory is currently not supported, but the code 
checking for this case expects to see a FileNotFoundCacheException.

The code doesn't throw FileNotFoundCacheException; instead, 
_fs.getParentOf returns null, which triggers a NPE when acquiring the 
file attributes.  This is apparent when browsing the root of the 
filesystem with the WebDAV door (via a web-browser)

18 May 2011 16:12:02 (PnfsManager) [door:WebDAV-discordia@dCacheDomain:1305727921627 WebDAV-discordia PnfsGetFileAttributes] Error while retriving file attributes: null
java.lang.NullPointerException: null
        at org.dcache.chimera.namespace.ChimeraNameSpaceProvider.getFileAttributes(ChimeraNameSpaceProvider.java:679) ~[dcache.jar:na]
        at org.dcache.chimera.namespace.ChimeraNameSpaceProvider.getFileAttributesForPermissionHandler(ChimeraNameSpaceProvider.java:671) ~[dcache.jar:na]
        at org.dcache.chimera.namespace.ChimeraNameSpaceProvider.getFileAttributes(ChimeraNameSpaceProvider.java:812) ~[dcache.jar:na]
        at diskCacheV111.namespace.PnfsManagerV3.getFileAttributes(PnfsManagerV3.java:1930) [dcache.jar:na]
        at diskCacheV111.namespace.PnfsManagerV3.processPnfsMessage(PnfsManagerV3.java:1720) [dcache.jar:na]
        at diskCacheV111.namespace.PnfsManagerV3$ProcessThread.run(PnfsManagerV3.java:1527) [dcache.jar:na]
        at java.lang.Thread.run(Thread.java:636) [na:1.6.0_17]
18 May 2011 16:12:02 (WebDAV-discordia) [door:WebDAV-discordia@dCacheDomain:1305727921627] Internal server error: org.dcache.webdav.WebDavException: java.lang.NullPointerException

The problem is not present when browsing other (i.e., non-root) 
directories.

This patch fixes the problem by checking whether getParentOf returns 
null, rather than trapping FileNotFoundCacheException.

Target: trunk
Patch: http://rb.dcache.org/r/3343/
Acked-by: Gerd Berhmann