info: add NDC describing the current data-gathering activity
The info service uses several data-gathering activities (DGA) that are 
scheduled to run at different rates and process asynchronously, as 
results are received.  Reply messages are processed to build a concise 
set of changes (adding or removing metrics) to the current state.  
These changes are processed by a thread dedicated to this task.

Currently, the logging contains no diagnostic context.  Any problem 
discovered when processing the changes is then difficult (almost 
impossible) to trace back to the initial DGA.  This makes diagnosing 
some problems hard.

This patch fixes this by adding the DGA's name to the nested 
diagnostic context.  The NDC is preserved when the update task is 
added to the queue of pending updates and finally run, so that 
diagnostic information is available if any problem is discovered.

I'm requesting a back-port this patch as it may prove very useful when 
diagnosing problems.

Target: trunk
Request: 1.9.12
Request: 1.9.11
Request: 1.9.10
Request: 1.9.5
Patch: http://rb.dcache.org/r/3383
Acked-by: Gerd Behrmann