cells: handle duplicate commands correctly
The admin command "set printout" is defined by the method 
ac_set_printout method and online help is provided as fh_set_printout 
and hh_set_printout fields.  These are discovered by reflection when a 
class is introduced to the CommandInterpreter object.

All information about a command is stored in a CommandEntry class.  A 
CommandEntry includes the Method and Field objects discovered by 
reflection.  The method and all fields discovered via reflection each 
have an associated object.  This is the object from which the 
method/fields were introspected from, and which is to be used when 
fetching information about this field or invoking the method.

There are two bugs in CommandInterpreter.

First, when a command is introspected that has the same name as a 
previously discovered command, discovered values are overwritten; 
however, the fields (full-help, help-hint, acls) are not cleared 
first.  If the new command has fewer fields then the previous declared 
command's fields are retained.

This is the case for the ac_set_printout method.  This command is 
initially registered via CellShell constructor calling superclass 
CommandInterpreter default constructor, which calls 
addCommandListener(this).  CellShell has ac_set_printout with 
full-help and help-hint fields.

Later, in SystemCell, there is this line:

       _cellShell.addCommandListener(this);

This registers all the commands in the SystemCell hierarchy, which 
includes those in CellAdapter.  CellAdapter includes an alternative 
set printout command, but defines only the full-help field; the 
help-hint field is missing.

Because the help-hint field is missing, when CommandInterpreter 
discovers the "set printout" command from CellAdapter, the help-hint 
from CellShell is kept.

The second bug is that, when evaluating the value of the help-hint 
field, the code uses the object from which the ac_set_printout method 
was discovered, not the object for the help-hint field.  Under normal 
circumstances, this distinction doesn't matter as it is the same 
object use for introspection with the method and the three fields 
(help-hint, full-help and acls).

However, with the first bug, the second bug results in the error 
message:

java.lang.IllegalArgumentException: Can not set java.lang.String field 
dmg.cells .nucleus.CellShell.hh_set_printout to 
dmg.cells.nucleus.SystemCell

Note that this is an attempt to extract the value of field 
"dmg.cells.nucleus.CellShell.hh_set_printout" from a SystemCell 
object.

This patch addresses this issue by clearing the three field-derived 
values when adding a command if that command has already been set.  
When adding the second ac_set_printout command; all fields are cleared 
so the phantom help-hint is removed.

The second bug (using the wrong object) is left to reduce the size of 
the patch.  A later patch should fix this issue, possibly by 
refactoring the code slightly.

Target: trunk
Request: 1.9.12
Request: 1.9.11
Request: 1.9.10
Request: 1.9.5
Patch: http://rb.dcache.org/r/3391/
Acked-by: Gerd Behrmann