info: fix "StateUpdate with zero updates encountered" message
Incoming messages to the info cell and converted into a set of changes 
to the info service's view of dCache's current state.  The mechanism 
that processes these updates checks whether an update has at least one 
update; if not, then the update is considered broken and it issues the 
warning "StateUpdate with zero updates encountered".

One of the data-gathering activities inside info queries the routing 
manager of the different domains to query the routing table.  This is 
to identify all the well-known cells within dCache.  The code that 
accepts the reply messages from the routing manager always requests an 
update to the info's state.

Unfortunately, the update generated from the reply from a domain 
without any well-known cells will not modify the state, so triggering 
the "StateUpdate with zero updates encountered" safety.

This patch fixes the problem: if the reply is from a domain without 
any well-known cells then this fact is logged (at debug level) and no 
update request is generated.

Target: trunk
Request: 1.9.5
Request: 1.9.10
Request: 1.9.11
Request: 1.9.12
Patch: http://rb.dcache.org/r/3385/
Acked-by: Gerd Behrmann