cells: re-enable logging for ssh
The SSH protocol implementation within cells includes a number of 
logging statements.  It appears that these statements would log to 
stdout in the past, but currently do nothing.  There are other 
statements that log errors but do so by printing directly to stderr.

This patch takes the existing logging information and uses a slf4j 
logger to provide the information.  The error/warning logging is 
adjusted so it goes through the logger, too.

Target: trunk
Patch: http://rb.dcache.org/r/3395/
Acked-by: Tigran Mkrtchyan