Add an automatic background checksum verification on pools
Background checksum verification (the scrubber) is implemented as a
per-pool thread that if enabled (default disabled) runs in the
background and checks all pool data every n-hour interval (default 24)
while also adjusting its checksum throughput to be below a certain
limit (default infinite MiB/s).

Whenever a checksum mismatch is detected the corresponding file is
marked as BROKEN and an error is logged to the target
org.dcache.pool.classic.ChecksumScanner. If there's an I/O error while
trying to verify the file, an error is logged and the file is skipped.

If a single scan or a full scan is started when the scrubber is
active, the scrubber will be suspended until the single/full scan has
finished.

The already existing 'csm set policy' stub arguments '-frequently' and
'-every' have been removed and replaced with '-scrub' and '-period'
respectively. A new argument '-limit' has been added for setting the
throughput limit. '-filedelay' has also been removed.

Target: trunk
Patch: http://rb.dcache.org/r/3006/
Acked-by: Gerd Behrmann
Acked-by: Tigran Mkrtchyan

DOC:
A section about the scrubber needs to be added. The reference to the
removed csm set policy argument "-frequently" should be removed.