gplazma: Fail authorization if storage-authzdb is missing a record
Several gPlazma 1 plugins rely on a combination of some mapping followed
by a storage-authzdb mapping. If the storage-authzdb file was missing an
entry for the user name generated by the first mapping then an error was
logged, however the authorization would succeed with UID 0 (essentially
giving unauthorized access to the system).

This patch changes the logic to fail the login.

The patch needs to be adjusted for some of the old branches.

Target: trunk
Request: 1.9.12
Request: 1.9.11
Request: 1.9.10
Request: 1.9.5
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3387/