xrootd: Reduce the negative effect of busy pools
The xrootd door had a rather aggressive timeout for instantiating an
xrootd mover (4 seconds). We have seen that this regularly is a problem
during peaks. Therefore this patch changes the default to 15 seconds.

The patch also add a message handler for PoolIoFileMessage. The message
handler is only invoked when the corresponding transfer timed out or was
interrupted. Therefore the message handler kills the mover to avoid that
it fills up a transfer slot.

Finally, the patch adds a mover kill in case the door or the client
times out waiting for a redirect message from the mover.

Target: trunk
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3403/