Use properties for gPlazma2 plugin configuration
gPlazma2 used string[] to pass parameters to the plugins. This way
either the order of the parameters has to fixed and therefore optional
parameters are hard to handle or keys and values have to be passed in
alternating order which is somewhat unintuitive or each string has to
contain a key/value pair in some special string format that has then
be parsed by each plugin which is not DRY at all.
Furthermore we need to get dCache-environment variables into the plugins
(e.g. ${DCACHE_ETC}) to automatically adjust default pathes, etc.
Both tasks can be easily addressed using java.util.Properties.

So with this patch all parameters to all plugins are passed as Properties.
that are created during configuration parsing.

Patch: http://rb.dcache.org/r/3255/
Acked-by: behrmann 
Target: trunk
Request: 1.9.12