srm: Compute file locality in the SRM rather than in pool manager
SRM for some operations needs to compute the file locality. It used to
query pool manager for this information. However all information to
compute the file locality is available to the pool monitor component.
Since the pool monitor can be serialized, we have the chance to move the
file locality computation from pool manager to the SRM itself.

This patch does exactly that: In moves the getFileLocality method into
PoolMonitorV5. It updated the SRM to periodically fetch the
PoolMonitorV5 object from the pool manager. To do so it uses the
existing storage info update loop which now no longer needs to fetch the
pool list from pool manager (since the pool list is embedded in the
PoolMonitorV5 instance). Finally, the patch updates the file locality
computation in the SRM to use the PoolMonitorV5 instance.

The PoolManagerGetFileLocality message is removed since it is no longer
used.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3414/