srm: Add option not to pin online files
Files with an access latency of ONLINE are always on disk. To reduce the
latency of srmPrepareToGet one may wish to skip pinning such files. In a
disk-only system one may also wish to completely disable the pin
manager.

This patch adds an option to skip pinning files with an access latency
of ONLINE.

Note that in dCache an ONLINE file may still have a file locality of
NEARLINE or UNAVAILABLE. This happens when the pool with the file is not
configured as a read pool for this particular request or if that pool is
down. Therefore the SRM computes the file locality of such files and if
NEARLINE it will ask pool manager for a read pool; this in turn triggers
a pool to pool transfer to a read pool.

Also note that disk only files have to have an access latency of ONLINE
for the new code to recognize them as disk files. Legacy systems that
use LSF mode precious to simulate tape with disks will still pin files.

Finally note that when enabled, files that are on both tape and disk
(permanently), ie CUSTODIAL ONLINE files, will not be pinned either.
However if the ONLINE copy is lost, a stage is triggered.

One issue that arises when skipping pinning is that the SRM code relies
on a pin id to be generated. Therefore this patch introduces the special
pin id 'disk', which indicates that pinning was skipped.

Another issue is that read availability now depends on the cached pool
monitor instance in the SRM. This is currently only updated every 5
minutes. Consequently it can happen that an upload succeeds, but a
download immediately following the upload fails because the SRM sees the
file as being unavailable. This should be addressed by increasing the
update frequency of the pool monitor. One may also add a rountrip to
pool manager when a file appears to be unavilable.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3417/