gplazma: Fix logging of VOMS plugin
PKIStore internally creates a thread to periodically reload the store.
Since threads inherit the MDC of the parent thread, controlling the MDC
at the point of thread creation is important.

VomsPlugin instantiates the PKIStore as a side effect the first time a
login request is processed. Currently this means that the MDC of the
request is associated with the reload thread. The consequence is that
the wrong MDC is logged whenever the PKIStore reloads itself.

Target: trunk
Request: 1.9.12
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3424/