srm: Don't cancel copy transfer if it already succeeded
The SRM copy implementation listens to state changes of the request:
When changed to a final state while having a transfer manager transfer
ID, the request submits a cancel message to the transfer manager. This
sounds reasonable, except that the transfer ID is never cleared in the
first place. Thus the following message appears in the log:

01 Jun 2011 09:49:00 (RemoteTransferManager) [v2:srmCopy:50415134:-2145571568:-2145571567 SRM-dcache-vm CancelTransfer] can not find handler with id=-9223372036831775230

This patch adds an additional check such that a cancel message is only
submitted when the request moves to a final state that is not Done.

Target: trunk
Request: 1.9.12
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3426/