gplazma: Move plugin defaults to gplazma.properties
The patch moves the plugin default values out of the Java code and into
the gplazma.properties file.

Properties are renamed to fit inside the dCache configuration property
scheme. The general format is gplazma.<plugin>.key. I chose keys that
are shorter than those we used originally in the plugins.

With this patch, plugin configuration values can be specified in
gplazma.conf, in the layout file, and in gplazma.conf.

I had to workaround a bug in cells/universal spring cell: The environment
is set *after* the constructor was called (obviously); since the cell
was initialized in the constructor, the environment was not available
during initialization. The current solution is an ugly workaround until
we find a better solution inside cells.

Release notes and documentation changes needed: gPlazma 2 has no public
documentation yet, however the internal drafts need to be updated.

Target: trunk
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3437/