ftp: Improve RFC 3659 compliance
RFC 3659 specifies the MLSD and MLST commands, among other things.
dCache implements these commands, but fails to follow a recommendation
in the spec:

   The file name returned in the MLST response should be the same name
   as was specified in the MLST command, or, where TVFS is supported, a
   fully qualified TVFS path naming the same file.  Where no argument
   was given to the MLST command, the server-PI may either include an
   empty file name in the response, or it may supply a name that refers
   to the current directory, if such a name is available.  Where TVFS is
   supported, a fully qualified pathname of the current directory SHOULD
   be returned.

In dCache we until now used the same output for MLST and MLSD, which
means that we only printed the file name. For MLSD we SHOULD either
print the request argument or the fully qualified path.

This patch improves compliance by outputting the fully qualified path
for MLST.

The patch adds TVFS as a feature in the FEAT output. The rest of the
door already implements TVFS, so we might as well advertize it (without
TVFS there is no such thing as a path in FTP; doing MLST /foo/bar is not
guaranteed to work as /foo/bar may not be a valid path in the server
OS's file system [eg on Windows]; TVFS defines a trivial hiearchical
name space using / as a separator).

Target: trunk
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3439/