ftp: Implement UNIX specific facts
RFC 3659 introduces a mechanism for providing platform specific facts in
the MLST and MLSD output. The RFC refers to an IANA registry for such
additional facts, however the registry is empty and dCache thus didn't
implement any additional facts.

The RFC does however give a number of examples for such platform
specific facts. Among other it provides a number of facts for POSIX
systems (here considered synonmous with UNIX). It does however appear as
if other projects proceeded and implemented those facts. Eg both Globus
gridftp and uberftp implement them.

dCache does implement a POSIX file system (which is *not* related to
what we often refer to as POSIX IO), thus the UNIX facts do apply to
dCache. Since they are used by system we wish to be compatible with,
this patch implements UNIX.owner, UNIX.group, and UNIX.mode in dCache.
The patch also adds type=OS.UNIX=slink as a type for symbolic links.

The current implementation reports the UID and GID for owner and group.
A future patch can update this to use the new reverse map
infrastructure.

Since the patch improves compatibility with other projects, I request
backporting it to 1.9.12.

Target: trunk
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3442/