admin: fix multiple problems with ssh connections
When a sysadmin logs out of the admin interface, the SSH engine 
"leaks" a socket.  The socket is left in a half-closed state.  It is 
possible that the number of logins results in the Java process 
running out of file descriptors.  When this happens, the admin is 
forced to restart the domain.  This problem is particularly acute if 
the site monitors dCache using scripts that interract via the admin 
interface.

The SSH protocol has the following flow when shutting down: server 
sends a CLOSE message to client, client disconnects from server 
(half-closing the connection), server then closes the other half.

Because of the Java socket API, the server must attempt to read a 
socket to block until the socket is closed.  The current code-base 
fails to do this, requiring higher levels to attempt to read, even 
though this read will never succeed.

This patch fixes the problem by making the close method of 
SshInputStream and SshOutputStream return only after the session has 
finished and the socket is closed.  This is achieved by moving the 
support for closing the session completely into the SshStreamEngine, 
as the close method.  This method tells the client that the session 
is over, closes the output part of the connection and waits for the 
client to acknowledge the end-of-session and disconnect.  Any 
unexpected packets are discarded.

The close method of the SshInputStream and SshOutputStream simply 
call the close method of the SshStreamEngine.

The patch also fixes the problem where calling the close method of 
SshInputStream twice results in an attempt to close the socket twice.  
This triggers a java.nio.channels.ClosedChannelException, which is 
logged.

The patch also removes the unnecessary explicit close of the engine 
in StreamObjectCell when the StreamLoginCell exits the main loop.  
This is unnecessary as the cell's cleanUp method ensures that the 
connection is always closed.

NOTE that this patch currently doesn't apply cleanly to non-trunk 
branches.  This is because the patch context includes slf4j logging 
lines that were committed to trunk as commit 15665.  However, the 
patch doesn't depend on the earlier patch; it is fairly easy to 
provide a version of the patch that will commit cleanly to non-trunk 
branches if the earlier (slf4j-logging) patch isn't back-ported.

Target: trunk
Request: 1.9.12
Request: 1.9.11
Request: 1.9.10
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6218
Acked-by: Gerd Behrmann
Acked-by: Tigran Mkrtchyan