info-provider: remove GlueSA.ACBR legacy value work-around
Two bugs in different software (lcg-infosites and SAM) didn't handle 
publishing a virtual organisation as an SA.ACBR attribute 
("GlueSAAccessControlBaseRule" in LDAP-speak) with the "vo:" prefix.  
This is due to legacy behaviour, where SA.ACBR attributes were 
publishes as just the VO name.

Various patches introduced and updated a worked-around for this issue 
by publishing each ACBR attribute twice: once with the "vo:"-prefix 
and once with just the name of the VO.  The most appropriate commits 
are:

    http://rb.dcache.org/r/983/   r.12697
    http://rb.dcache.org/r/1041/  r.12783

At the same time as the work-around were committed, the problems were 
reported in Savannah.

    https://savannah.cern.ch/bugs/?57787
    https://savannah.cern.ch/bugs/?57919 

From the Savannah entries, these bugs are now fixed and deployed in 
the production environment; therefore, the work-around is no longer 
needed.

More recently, the GLUE v1.3 checking service (gstat) is now marking 
storage-elements that publish ACBR values with the legacy value as 
having a critical problem.  This is clearly wrong and the gstat 
developers have agreed to reclassify this as a warning; however, there 
is a general movement to remove these legacy properties.  DPM 
info-provider has already done this.  The new WLCG Information Officer 
has OK-ed removing the values.

Therefore, this patch removes the work-around for 1.9.12 and trunk.  
A similar patch is available for releases prior to 1.9.12.

FOR RELEASE NOTES:

This release removes the GlueSAAccessControlBaseRule work-around that publishes legacy values for broken clients.  Any problems with lcg-infosites should be solved by upgrading to the latest version of the UI.

Target: trunk
Request: 1.9.12
Ticket: https://gus.fzk.de/ws/ticket_info.php?ticket=69943
Patch: http://rb.dcache.org/r/3453/
Acked-by: Gerd Behrmann