chimera cli: fix and update readtag and writetag, and add writefile
Chimera provides a CLI to allow operations on Chimera without 
starting the NFS server and mounting the filesystem.  One of these 
commands (writetag) accepts data only from STDIN.

That the writetag operation only accepts data from STDIN is a 
limitation.  For short tags, it may be desirable to write the tag 
data as a command-line argument.  Also, there is currently no way of 
writing data into a Chimera file without starting the NFS server and 
mounting the filesystem.  Configuring a dCache instance for use via 
URL-less dcap requires creating files in Chimera; so this limitation 
is awkward.

The current implementation of readtag is broken (the argument offset 
is wrong).  This patch fixes that problem.

This patch updates the writetag operation so it may take data from 
STDIN or as one of the command-line arguments.  It also removes the 
limitation on the size of the tag's data.

The patch also introduces the writefile operation that creates a 
file, if not already present, and writes the supplied content into 
that file.  The content of the file may be supplied either from 
STDIN or from a command-line argument (similar to writetag).

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3446
Acked-by: Gerd Berhmann