configuration: add missing httpd property with it's obsolete annotation
The httpdEnablePoolCollector is a property introduced to remove a 
commented-out part of the code that sites would uncomment if they chose 
to enable it.  This was for 1.9.5; since then, we've decided to always 
enable the PoolCollector, so the property is now obsolete.

For some reason, the property wasn't included in our list of default 
properties, so wasn't annotated as obsolete.  This leaves sites 
uncertain about the status of this property.

This patch fixes this.  It adds the property as obsolete, along with a 
description of why.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3458/
Acked-by: Tanja Baranova